+++ packages/importer/src/index.mts
@@ -1,2 +1,3 @@
 export type { ImportLocaleMapping, LocaleMapping } from '@typesafe-i18n/runtime/core.mjs'
-export { storeTranslationsToDisk, storeTranslationToDisk } from './importer'
+export { storeTranslationsToDisk, storeTranslationToDisk } from './importer.mjs'
+
